+++ providers/darwin/kernel_darwin.go
@@ -15,6 +15,8 @@
 // specific language governing permissions and limitations
 // under the License.
 
+// +build !386
+
 package darwin
 
 import (